+++ src/FlexiChains.jl
@@ -89,7 +89,7 @@ using MCMCDiagnosticTools: ess, rhat, mcse
 export ess, rhat, mcse
 # For maximum ease of use with Turing...
 const VNChain = FlexiChain{VarName}
-export VarName, @varname, VNChain, Prefixed
+export VarName, @varname, VNChain
 
 # Attempt to precompile _some_ stuff, especially for VarName. This cuts the TTFX by about
 # 2-3x.